Task 3
A restaurant menu UI/UX project is the process of designing a menu that is both visually appealing and easy to use. 
The aim was to design a good restaurant menu to make it easy for customers to find the dishes they want and to encourage them to order more.